backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-bigquery
google-cloud-storage
db-dtypes
duckdb
pandas
pyarrow
yfinance
lxml